Get BOM agriculture bulletin information for select stations

Fetch the BOM agricultural bulletin information.

get_ag_bulletin(state = "AUS")

Arguments

state

Australian state or territory as full name or postal code. Fuzzy string matching via base::agrep() is done. Defaults to "AUS" returning all state bulletins, see Details for more.

Value

A data frame as a weatherOz_tbl object (inherits and is fully compatible with data.table::data.table) of Australia BOM agricultural bulletin information.

Details

Allowed state and territory postal codes, only one state per request or all using 'AUS'.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

Note

Data and Information Use Please note the copyright notice and disclaimer, http://www.bom.gov.au/other/copyright.shtml related to the use of this information. Users of this information are deemed to have read and accepted the conditions described therein.

References

Agricultural observations are retrieved from the Australian Bureau of Meteorology (BOM) Weather Data Services Agriculture Bulletins,
http://www.bom.gov.au/catalogue/observations/about-agricultural.shtml.

And also,

Australian Bureau of Meteorology (BOM)) Weather Data Services Observation of Rainfall,
http://www.bom.gov.au/climate/how/observations/rain-measure.shtml.

Station location and other metadata are sourced from the Australian Bureau of Meteorology (BOM) webpage, Bureau of Meteorology Site Numbers:
http://www.bom.gov.au/climate/cdo/about/site-num.shtml.

See also

Author

Adam H. Sparks, adamhsparks@dpird.wa.gov.au and Paul Melloy paul@melloy.com.au

Examples

if (FALSE) { # interactive()

ag_bulletin <- get_ag_bulletin(state = "QLD")
ag_bulletin
}