Get a List of Available BOM Radar Imagery

Fetch a listing of available BOM radar imagery from ftp://ftp.bom.gov.au/anon/gen/radar/ to determine which files are currently available for download. The files available are the most recent radar imagery for each location, which are updated approximately every 6 to 10 minutes by the BOM. Ported from bomrang.

Usage

get_available_radar(radar_id = "all")

Arguments

radar_id

Numeric. BOM radar of interest for which a list of available images will be returned. Defaults to all images currently available.

Value

A data.table::data.table() of all selected radar locations with location information and product_ids.

Details

Valid BOM radar ID for each location required.

References

Australian Bureau of Meteorology (BOM) radar image https://www.bom.gov.au/weather-and-climate/rain-radar-and-weather-maps.

See also

Author

Dean Marchiori, deanmarchiori@gmail.com, and Adam H. Sparks, adamhsparks@gmail.com

Examples

if (FALSE) { # interactive()

# Check availability radar imagery for Wollongong (radar_id = 3)
imagery <- get_available_radar(radar_id = 3)

imagery
}