Get DataDrill Weather Data in the APSIM Format From SILO

Fetch APSIM .met file formatted weather data from the weather data from the SILO API of spatially interpolated weather data (DataDrill). The daily climate surfaces have been derived either by splining or kriging the observational data. The returned values contain “source” columns, which denote how the observations were derived. The grid spans 112° to 154°, -10° to -44° with resolution 0.05° latitude by 0.05° longitude (approximately 5 km × 5 km).

get_data_drill_apsim(
  longitude,
  latitude,
  start_date,
  end_date = Sys.Date(),
  api_key
)

Arguments

longitude

A single numeric value representing the longitude of the point-of-interest.

latitude

A single numeric value representing the latitude of the point-of-interest.

start_date

A character string or Date object representing the beginning of the range to query in the format “yyyy-mm-dd” (ISO8601). Data returned is inclusive of this range.

end_date

A character string or Date object representing the end of the range query in the format “yyyy-mm-dd” (ISO8601). Data returned is inclusive of this range. Defaults to the current system date.

api_key

A character string specifying a valid email address to use for the request. The query will return an error if a valid email address is not provided.

Value

An apsimx object of class ‘met’ with attributes.

Included Values

rain (mm)

Rainfall

maxt (˚C)

Maximum temperature

mint (˚C)

Minimum temperature

vp (hPa)

Vapour pressure

evap_pan (mm)

Class A pan evaporation

radiation (Mj/m1)

Solar exposure, consisting of both direct and diffuse components

Value information

Solar radiation: total incoming downward shortwave radiation on a horizontal surface, derived from estimates of cloud oktas and sunshine duration2.

Evaporation and evapotranspiration: an overview of the variables provided by SILO is available here, https://data.longpaddock.qld.gov.au/static/publications/Evapotranspiration_overview.pdf.

Data codes

Where the source code is a 6 digit string comprising the source code for the 6 variables. The single digit code for each variable is:

0

an actual observation;

1

an actual observation from a composite station;

2

a value interpolated from daily observations;

3

a value interpolated from daily observations using the anomaly interpolation method for CLIMARC data;

6

a synthetic pan value; or

7

an interpolated long term average.

References

Rayner, D. (2005). Australian synthetic daily Class A pan evaporation. Technical Report December 2005, Queensland Department of Natural Resources and Mines, Indooroopilly, Qld., Australia, 40 pp.

Morton, F. I. (1983). Operational estimates of areal evapotranspiration and their significance to the science and practice of hydrology, Journal of Hydrology, Volume 66, 1-76.

See also

Author

Rodrigo Pires, rodrigo.pires@dpird.wa.gov.au, and Adam Sparks, adam.sparks@dpird.wa.gov.au

Examples

if (FALSE) {
# requires an API key as your email address
# Source data from latitude and longitude coordinates (gridded data) for
# max and minimum temperature and rainfall for Southwood, QLD.
wd <- get_data_drill_apsim(
  latitude = -27.85,
  longitude = 150.05,
  start_date = "20221001",
  end_date = "20221201",
  api_key = "your@email"
)
}