Get minute weather data from DPIRD Weather 2.0 API

Get nicely formatted minute weather station data from the DPIRD weather station network over a maximum 24 hour period.

get_dpird_minute(
  station_code,
  start_date_time = lubridate::now() - lubridate::hours(24L),
  minutes = 1440L,
  which_values = "all",
  api_key
)

Arguments

station_code

A character string or vector of the DPIRD station code for the station of interest.

start_date_time

A character string representing the start date and time of the query in the format 'yyyy-mm-dd-hh-mm'. Defaults to 24 hours before the current local system time, returning the most recent 24 hour observations rounded to the nearest minute. This function does its best to decipher many date and time formats but prefers ISO8601.

minutes

An integer value that provides the number of observations to be returned. Defaults to 1440 minutes for 24 hours of observations.

which_values

A vector of weather values to query from the API. See Available Values section for valid available codes. Defaults to all available values, "all".

api_key

A character string containing your API key from DPIRD, https://www.agric.wa.gov.au/web-apis, for the DPIRD Weather 2.0 API.

Available Values:

all (includes all of the following),

airTemperature,

dateTime,

dewPoint,

rainfall,

relativeHumidity,

soilTemperature,

solarIrradiance,

wetBulb,

wind,

windAvgSpeed,

windMaxSpeed, and

windMinSpeed

Value

a data.table::data.table with 'station_code' and date interval queried together with the requested weather variables.

Note

Please note this function converts date-time columns from Coordinated Universal Time 'UTC' returned by the API to Australian Western Standard Time 'AWST'.

See also

Examples

if (FALSE) {
library(lubridate)

yesterday <- now() - hours(24)

get_dpird_minute(station_code = "SP",
                 start_date_time = "2018-02-01 13:00:00",
                 minutes = 1440,
                 api_key = YOUR_API_KEY,
                 which_values = c("airTemperature",
                                  "solarIrradiance",
                                  "wind"))
}