Get extreme weather event summaries for a single DPIRD station

Get extreme weather event summaries for a single DPIRD station

get_extreme_weather(
  station_code,
  which_values = "all",
  group = "rtd",
  include_closed = FALSE,
  api_key = NULL
)

Arguments

station_code

A character string with the station code for the station of interest.

which_values

A character string with the type of extreme weather to return. See Available Values for a full list of valid values. Defaults to 'all', returning the full list of values unless otherwise specified.

group

Filter the stations to a predefined group. These need to be supported on the back end; 'all' returns all stations, 'api' returns the default stations in use with the API, 'web' returns the list in use by the https:://weather.agric.wa.gov.au and 'rtd' returns stations with scientifically complete data sets. Defaults to 'rtd'.

include_closed

A Boolean value that defaults to FALSE. If set to TRUE the query returns closed and open stations. Closed stations are those that have been turned off and no longer report data. They may be useful for historical purposes.

api_key

A character string containing your API key from DPIRD, https://www.agric.wa.gov.au/web-apis, for the DPIRD Weather 2.0 API.

Available Values for which_values

all (returns all of the following values),

erosionCondition,

erosionConditionLast7Days,

erosionConditionLast7DaysDays,

erosionConditionLast7DaysMinutes,

erosionConditionLast14Days,

erosionConditionLast14DaysDays,

erosionConditionLast14DaysMinutes,

erosionConditionMonthToDate,

erosionConditionMonthToDateDays,

erosionConditionMonthToDateMinutes,

erosionConditionMonthToDateStartTime,

erosionConditionSince12AM,

erosionConditionSince12AMMinutes,

erosionConditionSince12AMStartTime,

erosionConditionYearToDate,

erosionConditionYearToDateDays,

erosionConditionYearToDateMinutes,

erosionConditionYearToDateStartTime,

frostCondition,

frostConditionLast7Days,

frostConditionLast7DaysDays,

frostConditionLast7DaysMinutes,

frostConditionLast14Days,

frostConditionLast14DaysDays,

frostConditionLast14DaysMinutes,

frostConditionMonthToDate,

frostConditionMonthToDateDays,

frostConditionMonthToDateMinutes,

frostConditionMonthToDateStartTime,

frostConditionSince9AM,

frostConditionSince9AMMinutes,

frostConditionSince9AMStartTime,

frostConditionTo9AM,

frostConditionTo9AMMinutes,

frostConditionTo9AMStartTime,

frostConditionYearToDate,

frostConditionYearToDate,

frostConditionYearToDateMinutes,

frostConditionYearToDateStartTime,

heatCondition,

heatConditionLast7Days,

heatConditionLast7DaysDays,

heatConditionLast7DaysMinutes,

heatConditionLast14Days,

heatConditionLast14DaysDays,

heatConditionLast14DaysMinutes,

heatConditionMonthToDate,

heatConditionMonthToDateDays,

heatConditionMonthToDateMinutes,

heatConditionMonthToDateStartTime,

heatConditionSince12AM,

heatConditionSince12AMMinutes,

heatConditionSince12AMStartTime,

heatConditionYearToDate,

heatConditionYearToDateDays,

heatConditionYearToDateMinutes, and

heatConditionYearToDateStartTime

Value

a data.table::data.table of one row with 'station_code', 'station_name', 'latitude', 'longitude', 'date_time' of the query and the extreme weather information according to the value(s) selected.

See also

Author

Rodrigo Pires, rodrigo.pires@dpird.wa.gov.au and Adam H. Sparks, adam.sparks@dpird.wa.gov.au.

Examples

if (FALSE) {
# You must have an DPIRD API key to proceed
my_key <- rstudioapi::askForSecret()

# Query Bonnie Rock station for wind erosion and heat extreme events.
my_station <- "BR"
my_events <- c("erosion", "heat")

output <- get_extreme_weather(
  station_code = my_station,
  type = my_events,
  api_key = my_key
)

# Query multiple stations for all extreme events
# Provide a list of station (as strings)
these_stations <- list("MN", "ES", "KARI", "NO", "KA002", "CO001", "MA002")

# Row bind output lists (one for each station)
# together with [data.table::rbindlist]
outputs <- lapply(these_stations,
                  get_extreme_weather,
                  type = "all",
                  api_key = my_key)

df <- data.table::rbindlist(outputs)
}