Get BOM daily précis forecast for select towns from BOM

Fetch the BOM daily précis forecast and return a data frame of the seven-day town forecasts for a specified state or territory. Ported from bomrang.

get_precis_forecast(state = "AUS")

Arguments

state

Australian state or territory as full name or postal code. Fuzzy string matching via base::agrep() is done. Defaults to “AUS” returning all state bulletins, see Details for more.

Value

A data.table::data.table() of an Australia BOM précis seven day forecasts for BOM selected towns.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata for towns come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00013.dbf

See also

Author

Adam H. Sparks, adamhsparks@dpird.wa.gov.au and Keith Pembleton, keith.pembleton@usq.edu.au and Paul Melloy, paul@melloy.com.au

Examples

if (FALSE) { # interactive()

# get the short forecast for Western Australia
get_precis_forecast(state = "WA")
}