Get BOM Satellite Imagery

Fetch BOM satellite GeoTIFF imagery from ftp://ftp.bom.gov.au/anon/gen/gms/ and return a terra SpatRaster S4 class (see [terra::rast()]) or stars S3 stars object of GeoTIFF files. Files are available at ten minutes update frequency with a 24-hour delete time. It is suggested to check file availability first by using get_available_imagery(). Ported from bomrang with modifications.

Usage

get_satellite_imagery(product_id, scans = 1, compat = "terra")

Arguments

product_id

Character. BOM product ID to download and import as a terra SpatRaster S4 class (see [terra::rast()]) or stars S3 stars class object. A vector of values from get_available_imagery() may be used here. Value is required.

scans

Integer. Number of scans to download, starting with most recent and progressing backwards, e.g., 1 - the most recent single scan available , 6 - the most recent hour available, 12 - the most recent 2 hours available, etc. Negating will return the oldest files first. Defaults to 1. Value is optional.

compat

Character. A string indicating the R package with which the returned imagery should be formatted for use, one of terra or stars. Defaults to terra.

Value

A terra SpatRaster S4 class (see [terra::rast()]) or stars S3 stars class object as selected by the user by specifying compat of GeoTIFF images with layers named by BOM product ID, timestamp and band.

Details

Valid BOM satellite Product IDs for use with product_id include:

IDE00420

AHI cloud cover only 2km FD GEOS GIS

IDE00421

AHI IR (Ch13) greyscale 2km FD GEOS GIS

IDE00422

AHI VIS (Ch3) greyscale 2km FD GEOS GIS

IDE00423

AHI IR (Ch13) Zehr 2km FD GEOS GIS

IDE00425

AHI VIS (true colour) / IR (Ch13 greyscale) composite 1km FD GEOS GIS

IDE00426

AHI VIS (true colour) / IR (Ch13 greyscale) composite 2km FD GEOS GIS

IDE00427

AHI WV (Ch8) 2km FD GEOS GIS

IDE00430

AHI cloud cover only 2km AUS equirect. GIS

IDE00431

AHI IR (Ch13) greyscale 2km AUS equirect. GIS

IDE00432

AHI VIS (Ch3) greyscale 2km AUS equirect. GIS

IDE00433

AHI IR (Ch13) Zehr 2km AUS equirect. GIS

IDE00435

AHI VIS (true colour) / IR (Ch13 greyscale) composite 1km AUS equirect. GIS

IDE00436

AHI VIS (true colour) / IR (Ch13 greyscale) composite 2km AUS equirect. GIS

IDE00437

AHI WV (Ch8) 2km AUS equirect. GIS

IDE00439

AHI VIS (Ch3) greyscale 0.5km AUS equirect. GIS

Note

The original bomrang version of this function supported local file caching using hoardr. This version does not support this functionality any longer due to issues with CRAN and hoardr.

References

Australian Bureau of Meteorology (BOM) high-definition satellite images
https://www.bom.gov.au/australia/satellite/index.shtml.

See also

Author

Adam H. Sparks, adamhsparks@gmail.com

Examples

if (FALSE) { # interactive()
# Fetch AHI VIS (true colour) / IR (Ch13 greyscale) composite 1km FD
# GEOS GIS {terra} `SpatRaster`` object for most recent single scan
 available

imagery <- get_satellite_imagery(product_id = "IDE00425", scans = 1)
plot(imagery)

# Get a list of available image files and use that to specify files for
# download, downloading the two most recent files available

avail <- get_available_imagery(product_id = "IDE00425")
imagery <- get_satellite_imagery(product_id = avail, scans = 2)
plot(imagery)
}