Fetch a list with all available weather stations for the DPIRD Weather API or stations present in the SILO API.

Queries DPIRD and SILO APIs. If querying the DPIRD Weather API, the function return a data frame with metadata for the weather stations in the DPIRD network. Data providers include: DPIRD, DBCA, DFES, Water Corporation, Harvey Water and Pardoo Beef Corporation. If querying the SILO API, the function also return a data frame with metadata for the SILO weather stations network.

get_station_list(api = "weather", state = "wa", limit = 1000, api_key = NULL)

Arguments

api

name of the API to be queried. Needs to match to one of the existing options (currently 'weather' or 'silo'). Defaults to the "weather" API.

state

a string limiting the query to one or several states, defaults to "wa". Accepts "all", "wa", "sa", "nsw", "vic", "qld", "tas", "nt".

limit

a numeric value, limiting the amount of rows to return. Defaults to 1000.

api_key

User's API key from DPIRD (https://www.agric.wa.gov.au/web-apis)

Value

a data.table with station code, name, latitude, longitude, owner and state plus metadata details if any available.

Note

You can request your own API key from DPIRD for free by filling out the form found at https://www.agric.wa.gov.au/web-apis.

Author

Rodrigo Pires, rodrigo.pires@dpird.wa.gov.au

Examples

if (FALSE) { # interactive()
# Query DPIRD API
# You must have an api key to query the DPIRD API
mykey <- "YOUR_API_KEY"

dpird_stations <- get_station_list(api = "weather",
                                   state = "wa",
                                   api_key = mykey)

# Query the SILO API
silo_stations <- get_station_list(api = "silo",
                                   state = "nsw",
                                   api_key = mykey)

# Query the SILO API for multiple states
silo_stations_multi_state <- get_station_list(api = "silo",
                                              state = c("nsw", "nt"),
                                              api_key = mykey)
}