Get the latest DPIRD and SILO weather station metadata

Download the latest station locations and metadata for stations in the SILO and DPIRD APIs.

get_station_metadata(which_api = "all", api_key, status = FALSE, rich = FALSE)

Arguments

which_api

A string value that indicates which API to use. Valid values are 'all', for both SILO (BOM data) and DPIRD APIs; 'silo' for only stations from the SILO API (BOM data); or 'dpird' for stations from the DPIRD Weather 2.0 API. Defaults to "all".

api_key

A character string containing your API key from DPIRD, https://www.agric.wa.gov.au/web-apis, for the DPIRD Weather 2.0 API.

status

A Boolean string indicating whether to include closed stations' metadata. Use TRUE to include. Defaults to FALSE.

rich

A Boolean string indicating whether to return rich information about DPIRD's weather station(s), this does not affect the SILO stations' metadata, the variables for these observations will be NA. Defaults to FALSE.

Value

a data.table::data.table of BOM weather stations' metadata for stations available from SILO and weather stations' metadata for stations available from DPIRD's Weather 2.0 API with the following columns sorted by 'state' and 'station_name'.

station_code:	Unique station code. factor

station_name:	Unique station name. character

start:	Date observations start. date

end:	Date observations end. date

latitude:	Latitude in decimal degrees. numeric

longitude:	Longitude in decimal degrees. numeric

state:	State in which the station is located. character

elev_m:	Station elevation in metres. numeric

source:	Organisation responsible for the data or station maintenance. character

status:	Station status, one of 'open' or 'closed'. character

wmo:	World Meteorological Organisation, (WMO), number if applicable. numeric

rich values

capabilities:	a list of the station's capabilities (data that it records). character

probe_height:	temperature probe height in metres. double

rain_gauge_height	rain gauge height in metres. double

wind_probe_heights:	wind probe heights always 3 metres, although some have 10 metre probes. integer

Note

For stations in the SILO API, BOM does not report the exact date on which stations opened or closed, only the year. Therefore the 'start' and 'end' columns will indicate January 1 of the year that a station opened or closed, whereas stations in the DPIRD network have the date to the day. For BOM stations that are closed for the current year, this indicates that the station closed sometime during the current year prior to the request being made. NA in the current year indicates a station is still open.

References

Station location and other metadata are sourced from the Australian Bureau of Meteorology (BOM) webpage, Bureau of Meteorology Site Numbers:
http://www.bom.gov.au/climate/cdo/about/site-num.shtml and http://www.bom.gov.au/climate/data/lists_by_element/stations.txt and the DPIRD Weather 2.0 API.

Author

Adam H. Sparks, adam.sparks@dpird.wa.gov.au

Examples

if (FALSE) { # interactive()
# fetch SILO metadata
get_station_metadata()
}