Western Australia Southwest Agriculture Region Geospatial Polygon

An sf object of the the WA South West Agricultural Region.

Usage

south_west_agricultural_region

Format

An sf::sf() polygon object

Source

Western Australia Department of Primary Industries and Regional Development under a Creative Commons Attribution 4.0 Licence

Details

The zone managed for intensive agricultural activities in South-Western Australia. Also known as the South West Agricultural Area or Clearing Line. This zone defines the easternmost extent of land cleared for agricultural purposes.

Base data sets

Western Australian Land Information Authority - Captured from photographic interpretation of best available orthophotography at date of capture, dates range between 2007 and 2010.

Scale of capture

1:20,000

Coordinate Reference System

EPSG:4326 - WGS 84 – WGS84 - World Geodetic System 1984, used in GPS https://epsg.io/4326