Update internal database with latest BOM forecast towns

Download the latest select forecast towns from the BOM server and update internal database of précis forecast town names and AAC codes used by get_precis_forecast(). There is no need to use this unless you know that a forecast town exists in a more current version of the BOM précis forecast town name database that is not available in the database distributed with weatherOz. In fact, for reproducibility purposes, users are discouraged from using this function. Ported from bomrang.

update_forecast_towns()

Value

An updated internal package database of BOM précis forecast towns

References

Data are sourced from: Australian Bureau of Meteorology (BOM) webpage, “Weather Data Services”, http://www.bom.gov.au/catalogue/data-feeds.shtml.

See also

Author

Adam H. Sparks, adam.sparks@dpird.wa.gov.au

Examples

if (FALSE) { # interactive()

update_forecast_towns()
}