weatherOz: An API Client for Australian Weather and Climate Data Resources

Provides automated downloading, parsing and formatting of weather data for Australia through API endpoints provided by the Department of Primary Industries and Regional Development (DPIRD) of Western Australia and by the Science and Technology Division of the Queensland Government's Department of Environment and Science (DES). As well as the Bureau of Meteorology (BOM) of the Australian government precis and coastal forecasts, and downloading and importing radar and satellite imagery files. DPIRD weather data are accessed through public APIs provided by DPIRD, https://www.dpird.wa.gov.au/online-tools/apis/, providing access to weather station data from the DPIRD weather station network. Australia-wide weather data are based on data from the Australian Bureau of Meteorology (BOM) data and accessed through SILO (Scientific Information for Land Owners) Jeffrey et al. (2001) doi:10.1016/S1364-8152(01)00008-1 . DPIRD data are made available under a Creative Commons Attribution 3.0 Licence (CC BY 3.0 AU) license https://creativecommons.org/licenses/by/3.0/au/deed.en. SILO data are released under a Creative Commons Attribution 4.0 International licence (CC BY 4.0) https://creativecommons.org/licenses/by/4.0/. BOM data are (c) Australian Government Bureau of Meteorology and released under a Creative Commons (CC) Attribution 3.0 licence or Public Access Licence (PAL) as appropriate, see https://www.bom.gov.au/copyright for further details.

See also

Author

Maintainer: Rodrigo Pires rodrigo.pires@dpird.wa.gov.au (ORCID)

Authors:

Anna Hepworth anna.hepworth@gmail.com (ORCID)

Rebecca O'Leary bec.oleary@curtin.edu.au

Jonathan Carroll rpkg@jcarroll.com.au (ORCID)

James Goldie me@jamesgoldie.dev (ORCID)

Dean Marchiori deanmarchiori@gmail.com (ORCID)

Paul Melloy paul@melloy.com.au (ORCID)

Mark Padgham mark.padgham@email.com (ORCID)

Hugh Parsonage hugh.parsonage@gmail.com (ORCID)

Adam H. Sparks adamhsparks@gmail.com (ORCID)

Other contributors:

Keith Pembleton keith.pembleton@unisq.edu.au (ORCID) (Contributed code and ideas for original bomrang package that was used in the creation of weatherOz.) [contributor]

Maëlle Salmon maelle.salmon@yahoo.se (ORCID) (Contributed to debugging a nasty little bug with CI where timezones caused tests to fail due to vcr not recognising the URL when run outside of Australia-Perth TZ! Suggested the use of local_timzeone().) [contributor]

Max Moldovan max.moldovan@adelaide.edu.au (ORCID) (Contributed valuable feedback on package usage leading to improvements in the package structure and functionality.) [contributor]

Stephen Bradshaw stephen.bradshaw@dpird.wa.gov.au (ORCID) (Identified data structure issues and provided insights that improved package data quality and usability.) [contributor]

Jimmy Ng jimmy.ng@dpird.wa.gov.au [contributor]

Steve Collins steve.collins@dpird.wa.gov.au (Designed the hex logo for 'weatherOz' hex logo.) [contributor]

Laurens Geffert laurensgeffert@gmail.com [reviewer]

Sam Rogers sam.rogers@adelaide.edu.au [reviewer]

Western Australia Agriculture Authority (WAAA) [copyright holder]

Curtin University [copyright holder]