Get a list of nearby weather stations from coordinates or station ID code

Get a list of nearby weather stations from coordinates or station ID code

find_nearby_stations(
  latitude = NULL,
  longitude = NULL,
  site = NULL,
  distance_km = 100,
  api_key = NULL,
  wa_only = TRUE,
  dpird_only = FALSE
)

Arguments

latitude

A numeric value for latitude expressed as decimal degrees (DD) (WGS84).

longitude

A numeric value for longitude expressed as decimal degrees (DD) (WGS84).

site

A string with the station ID code for the station of interest. Optional and defaults to NULL.

distance_km

A numeric value for distance to limit the search from the station or location of interest. Defaults to 100 km.

api_key

A string value that is the user's API key from DPIRD (see https://www.agric.wa.gov.au/web-apis).

wa_only

Boolean, return stations in Western Australia only. Defaults to TRUE.

dpird_only

Boolean, return only DPIRD owned stations. Defaults to FALSE.

Value

a data.table with 'stationCode', 'stationName', 'latitude', 'longitude', 'state', 'owner' and 'distance'. Data are sorted by increasing distance from station or location of interest.

Note

You can request your own API key from DPIRD for free by filling out the form found at https://www.agric.wa.gov.au/web-apis.

Author

Rodrigo Pires, rodrigo.pires@dpird.wa.gov.au

Examples

if (FALSE) { # interactive()
# You must have an DPIRD API key to proceed
my_key <- rstudioapi::askForSecret()

# Query WA only stations and return both DPIRD's and BOM's stations for
# the Northam WA station, returning stations with 50 km of this station
wa_stn <- find_nearby_stations(
  site = "NO",
  distance_km = 50,
  api_key = "YOUR API KEY",
  dpird_only = FALSE,
  wa_only = TRUE
)

# Query Wagga Wagga BOM station.
wagga_stn <- find_nearby_stations(
  latitude = -35.1583,
  longitude = 147.4575,
  distance_km = 50,
  api_key = "YOUR API KEY",
  dpird_only = FALSE,
  wa_only = FALSE
)
}