Get DPIRD weather station availability metadata

Get the availability metadata of weather stations in the DPIRD weather station network.

get_dpird_availability(
  station_code = NULL,
  start_date = NULL,
  end_date = NULL,
  which_values = "availability",
  api_key
)

Arguments

station_code

A character string of the DPIRD station code for the station of interest. Defaults to NULL, returning metadata for all stations during the requested start_date and end_date interval.

start_date

A character string representing the beginning of the range to query in the format 'yyyy-mm-dd' (ISO8601). Will return data inclusive of this range. Defaults to NULL, returning data for the current year-to-date. Must be sent along with an end_date.

end_date

A character string representing the end of the range query in the format 'yyyy-mm-dd' (ISO8601). Will return data inclusive of this range. Defaults to NULL, returning data for the current year-to-date. Must be sent with a start_date.

which_values

A character string with the type of availability metadata to return. See Available Values for a full list of valid values. Defaults to 'availability', returning metadata for all stations.

api_key

A character string containing your API key from DPIRD, https://www.agric.wa.gov.au/web-apis, for the DPIRD Weather 2.0 API.

Available Values for which_values:

availability (which will return all of the following values),

availabilityCurrentHour,

availabilityLast7DaysSince9AM,

availabilityLast7DaysSince12AM,

availabilityLast14DaysSince9AM,

availabilityLast14DaysSince12AM,

availabilityLast24Hours,

availabilityMonthToDateSince12AM,

availabilityMonthToDateTo9AM,

availabilitySince9AM,

availabilitySince12AM,

availabilityTo9AM,

availabilityYearToDateSince12AM, and

availabilityYearToDateTo9AM

Value

a data.table::data.table with 'station_code' and the requested metadata.

Examples

if (FALSE) {
# You must have a DPIRD API key to proceed
# Use default for end data (current system date)
output <- get_dpird_availability(
            station_code = "BI",
            start_date = "20170101",
            end_date = "20171231",
            api_key = "YOUR API KEY")
}