Get the latest DPIRD and SILO weather station metadata

Download the latest station locations and metadata for stations in the SILO and DPIRD data sets.

get_station_metadata(api_key = NULL, which_api = "silo")

Arguments

api_key

A character string containing your API key from DPIRD, https://www.agric.wa.gov.au/web-apis, for the DPIRD weather API.

which_api

A string value that indicates which API to use. Defaults to "silo". Valid values are "all", for both SILO (BOM) and DPIRD weather station networks; "silo" for only stations in the SILO network; or "dpird" for stations in the DPIRD network.

@note BOM does not report the exact date on which stations opened or closed, only the year. Therefore the 'start' and 'end' columns will indicate January 1 of the year that a station opened or closed for BOM stations whereas stations in the DPIRD network have the date to the day. For BOM stations that are closed for the current year, this indicates that the station closed sometime during the current year prior to the request being made.

Value

a data.table of BOM weather stations' metadata for stations available from SILO with the following columns.

station_code:	Unique station ID. factor

station_name:	Unique station name. character

start:	Date observations start. date

end:	Date observations end. date

lat:	Latitude in decimal degrees. numeric

lon:	Longitude in decimal degrees. numeric

state:	State in which the station is located. character

elev.m:	Station elevation in metres. numeric

source:	Organisation responsible for the data or station maintenance. character

status:	Station status, one of "open" or "closed". character

wmo:	World Meteorological Organisation, (acronymWMO), number if applicable. numeric

Details

If ASGS.foyer is installed locally, this function will automatically check and correct any invalid state values for stations located in Australia in the SILO data.

ASGS.foyer must be installed to use this.

References

Station location and other metadata are sourced from the Australian Bureau of Meteorology (BOM) webpage, Bureau of Meteorology Site Numbers:
http://www.bom.gov.au/climate/cdo/about/site-num.shtml and http://www.bom.gov.au/climate/data/lists_by_element/stations.txt and DPIRD Weather 2.0 API.

Author

Adam H. Sparks, adam.sparks@dpird.wa.gov.au

Examples

if (FALSE) { # interactive()
# fetch SILO metadata
get_station_metadata()
}