weatherOz: An API Client for Australian Weather and Climate Data Resources

Provides automated downloading, parsing, cleaning, and formatting of weather data for Australia through API endpoints provided by the Department of Primary Industries and Regional Development ('DPIRD' ) of Western Australia and by the Science and Technology Division of the Queensland Government's Department of Environment and Science ('DES'). As well as Australian Government Bureau of Meteorology ('BOM') precis and coastal forecasts, agriculture bulletin data, and downloading and importing radar and satellite imagery files. 'DPIRD' weather data are accessed through public APIs provided by 'DPIRD', https://www.agric.wa.gov.au/weather-api-20, providing access to weather station data from 'DPIRD's' weather station network. Australia-wide weather data are based on data from the Australian Bureau of Meteorology (BOM) data and accessed through SILO (Scientific Information for Land Owners) (Jeffery et al., 2001, doi: 10.1016/S1364-8152(01)00008-1 ). 'DPIRD' data are made available under a Creative Commons Attribution 3.0 Licence (CC BY 3.0 AU) license https://creativecommons.org/licenses/by/3.0/au/deed.en. 'SILO' data are released under a Creative Commons Attribution 4.0 International licence (CC BY 4.0) https://creativecommons.org/licenses/by/4.0/. 'BOM' data are (c) Australian Government Bureau of Meteorology and released under a Creative Commons (CC) Attribution 3.0 licence or Public Access Licence (PAL) as appropriate, see http://www.bom.gov.au/other/copyright.shtml for further details.

Author

Maintainer: Rodrigo Pires rodrigo.pires@dpird.wa.gov.au (ORCID)

Authors:

Anna Hepworth anna.hepworth@dpird.wa.gov.au (ORCID)

Rebecca O'Leary rebecca.o'leary@agric.wa.gov.au

Jonathan Carroll rpkg@jcarroll.com.au (ORCID)

James Goldie me@jamesgoldie.dev (ORCID)

Dean Marchiori deanmarchiori@gmail.com

Paul Melloy paul@melloy.com.au (ORCID)

Mark Padgham mark.padgham@email.com (ORCID) [reviewer]

Hugh Parsonage hugh.parsonage@gmail.com (ORCID)

Keith Pembleton keith.pembleton@usq.edu.au (ORCID)

Adam H. Sparks adam.sparks@dpird.wa.gov.au (ORCID)

Other contributors:

Western Australia Agriculture Authority (WAAA) [copyright holder]